cleaned up delete button
